* Reshaped data sent to SVM such that all training data is sent at once now.  Previously I wasn't sure how this had to work, but it seems that training cannot be performed 'online'.  * Modified main.cpp to read all input images into the detector, then perform training, then classify a few images.  * Resized all test training images to 64x128 (so they are all now the same size).  * FIX: Added some missing conversions to 32bit float  * FIX: I was stupidly doing ones-based indexing into Mat arrays (somehow 'Mat' array makes me think of Matlab-style 'mxArrays', which use 1s based indexes).
Currently the program is capable of running and classifying images.

+++ src/Cell.cpp
@@ -16,8 +16,8 @@ void Cell::addImage( const cv::Mat& src )
 {
     mSource = src.clone();
 
-    Mat gradient_horz( mSource.size(), CV_64F );
-    Mat gradient_vert( mSource.size(), CV_64F );
+    Mat gradient_horz( mSource.size(), CV_32F );
+    Mat gradient_vert( mSource.size(), CV_32F );
     {
         Mat kernel_horz = ( Mat_< float >( 1, 3 ) << -1, 0, 1 );
         Mat kernel_vert = ( Mat_< float >( 3, 1 ) << -1, 0, 1 );
